引入 Redis 数据库和实现 Token 鉴权的开发

+++ src/Control.cpp
@@ -12,6 +12,9 @@
 
 using namespace std;
 
+// 无权限
+static Json::Value NoPermission;
+
 // --------------------用户------------------------
 Json::Value Control::RegisterUser(Json::Value &registerjson) {
     return UserList::GetInstance()->RegisterUser(registerjson);
@@ -30,23 +33,38 @@ Json::Value Control::SelectUserInfo(Json::Value &queryjson) {
 }
 
 Json::Value Control::UpdateUserInfo(Json::Value &updatejson) {
+    // 如果不是本人或者管理员 无权修改
+    if (!UserList::GetInstance()->IsAuthor(updatejson))
+        return NoPermission;
     return UserList::GetInstance()->UpdateUserInfo(updatejson);
 }
 
 Json::Value Control::SelectUserUpdateInfo(Json::Value &queryjson) {
+    // 如果不是本人或者管理员 无权修改
+    if (!UserList::GetInstance()->IsAuthor(queryjson))
+        return NoPermission;
     return UserList::GetInstance()->SelectUserUpdateInfo(queryjson);
 }
 
 Json::Value Control::SelectUserSetInfo(Json::Value &queryjson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
     return UserList::GetInstance()->SelectUserSetInfo(queryjson);
 }
 
 Json::Value Control::DeleteUser(Json::Value &deletejson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(deletejson))
+        return NoPermission;
     return UserList::GetInstance()->DeleteUser(deletejson);
 }
 
 // --------------------题目------------------------
 Json::Value Control::SelectProblemInfoByAdmin(Json::Value &queryjson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
     return ProblemList::GetInstance()->SelectProblemInfoByAdmin(queryjson);
 }
 
@@ -55,6 +73,9 @@ Json::Value Control::SelectProblem(Json::Value &queryjson) {
 }
 
 Json::Value Control::EditProblem(Json::Value &insertjson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(insertjson))
+        return NoPermission;
     Json::Value resjson;
     if (insertjson["EditType"].asString() == "Insert") {
         resjson = ProblemList::GetInstance()->InsertProblem(insertjson);
@@ -66,6 +87,9 @@ Json::Value Control::EditProblem(Json::Value &insertjson) {
 }
 
 Json::Value Control::DeleteProblem(Json::Value &deletejson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(deletejson))
+        return NoPermission;
     return ProblemList::GetInstance()->DeleteProblem(deletejson);
 }
 
@@ -74,6 +98,9 @@ Json::Value Control::SelectProblemList(Json::Value &queryjson) {
 }
 
 Json::Value Control::SelectProblemListByAdmin(Json::Value &queryjson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
     return ProblemList::GetInstance()->SelectProblemListByAdmin(queryjson);
 }
 
@@ -90,6 +117,9 @@ Json::Value Control::SelectAnnouncementList(Json::Value &queryjson) {
 }
 
 Json::Value Control::SelectAnnouncementListByAdmin(Json::Value &queryjson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
     return AnnouncementList::GetInstance()->SelectAnnouncementListByAdmin(queryjson);
 }
 
@@ -98,18 +128,30 @@ Json::Value Control::SelectAnnouncement(Json::Value &queryjson) {
 }
 
 Json::Value Control::SelectAnnouncementByEdit(Json::Value &queryjson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
     return AnnouncementList::GetInstance()->SelectAnnouncementByEdit(queryjson);
 }
 
 Json::Value Control::InsertAnnouncement(Json::Value &insertjson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(insertjson))
+        return NoPermission;
     return AnnouncementList::GetInstance()->InsertAnnouncement(insertjson);
 }
 
 Json::Value Control::UpdateAnnouncement(Json::Value &updatejson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(updatejson))
+        return NoPermission;
     return AnnouncementList::GetInstance()->UpdateAnnouncement(updatejson);
 }
 
 Json::Value Control::DeleteAnnouncement(Json::Value &deletejson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(deletejson))
+        return NoPermission;
     Json::Value resjson = AnnouncementList::GetInstance()->DeleteAnnouncement(deletejson);
 
     // 当评论模块完成时，将下面注释去掉
@@ -127,6 +169,9 @@ Json::Value Control::SelectDiscussList(Json::Value &queryjson) {
 }
 
 Json::Value Control::SelectDiscussListByAdmin(Json::Value &queryjson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
     return DiscussList::GetInstance()->SelectDiscussListByAdmin(queryjson);
 }
 
@@ -135,18 +180,30 @@ Json::Value Control::SelectDiscuss(Json::Value &queryjson) {
 }
 
 Json::Value Control::SelectDiscussByEdit(Json::Value &queryjson) {
+    // 如果不是用户，无权限
+    if (!UserList::GetInstance()->IsOrdinaryUser(queryjson))
+        return NoPermission;
     return DiscussList::GetInstance()->SelectDiscussByEdit(queryjson);
 }
 
 Json::Value Control::InsertDiscuss(Json::Value &insertjson) {
+    // 如果不是用户，无权限
+    if (!UserList::GetInstance()->IsOrdinaryUser(insertjson))
+        return NoPermission;
     return DiscussList::GetInstance()->InsertDiscuss(insertjson);
 }
 
 Json::Value Control::UpdateDiscuss(Json::Value &updatejson) {
+    // 如果不是本人或者管理员 无权修改
+    if (!UserList::GetInstance()->IsAuthor(updatejson))
+        return NoPermission;
     return DiscussList::GetInstance()->UpdateDiscuss(updatejson);
 }
 
 Json::Value Control::DeleteDiscuss(Json::Value &deletejson) {
+    // 如果不是作者或者管理员，无权限
+    if (!UserList::GetInstance()->IsAuthor(deletejson))
+        return NoPermission;
     Json::Value resjson = DiscussList::GetInstance()->DeleteDiscuss(deletejson);
 
     // 当评论模块完成时，将下面注释去掉
@@ -164,6 +221,9 @@ Json::Value Control::SelectSolutionList(Json::Value &queryjson) {
 }
 
 Json::Value Control::SelectSolutionListByAdmin(Json::Value &queryjson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
     return SolutionList::GetInstance()->SelectSolutionListByAdmin(queryjson);
 }
 
@@ -172,18 +232,30 @@ Json::Value Control::SelectSolution(Json::Value &queryjson) {
 }
 
 Json::Value Control::SelectSolutionByEdit(Json::Value &queryjson) {
+    // 如果不是用户，无权限
+    if (!UserList::GetInstance()->IsOrdinaryUser(queryjson))
+        return NoPermission;
     return SolutionList::GetInstance()->SelectSolutionByEdit(queryjson);
 }
 
 Json::Value Control::InsertSolution(Json::Value &insertjson) {
+    // 如果不是用户，无权限
+    if (!UserList::GetInstance()->IsOrdinaryUser(insertjson))
+        return NoPermission;
     return SolutionList::GetInstance()->InsertSolution(insertjson);
 }
 
 Json::Value Control::UpdateSolution(Json::Value &updatejson) {
+    // 如果不是作者或者管理员，无权限
+    if (!UserList::GetInstance()->IsAuthor(updatejson))
+        return NoPermission;
     return SolutionList::GetInstance()->UpdateSolution(updatejson);
 }
 
 Json::Value Control::DeleteSolution(Json::Value &deletejson) {
+    // 如果不是作者或者管理员，无权限
+    if (!UserList::GetInstance()->IsAuthor(deletejson))
+        return NoPermission;
     Json::Value resjson = SolutionList::GetInstance()->DeleteSolution(deletejson);
 
     // 当评论模块完成时，将下面注释去掉
@@ -197,6 +269,9 @@ Json::Value Control::DeleteSolution(Json::Value &deletejson) {
 
 // ----------------------评论----------------------------
 Json::Value Control::SelectCommentListByAdmin(Json::Value &queryjson) {
+    // 如果不是管理员，无权查询
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
     return CommentList::GetInstance()->SelectCommentListByAdmin(queryjson);
 }
 
@@ -209,6 +284,9 @@ Json::Value Control::GetComment(Json::Value &queryjson) {
 }
 
 Json::Value Control::InsertComment(Json::Value &insertjson) {
+    // 如果不是用户，无权限
+    if (!UserList::GetInstance()->IsOrdinaryUser(insertjson))
+        return NoPermission;
     // 文章添加评论数
     Json::Value updatejson;
     updatejson["ArticleId"] = insertjson["ArticleId"];
@@ -233,6 +311,9 @@ Json::Value Control::InsertComment(Json::Value &insertjson) {
 
 // Json(ArticleId,CommentId)
 Json::Value Control::DeleteComment(Json::Value &deletejson) {
+    // 如果不是用户，无权限
+    if (!UserList::GetInstance()->IsOrdinaryUser(deletejson))
+        return NoPermission;
     string articleid = deletejson["ArticleId"].asString();
 
     Json::Value resjson;
@@ -275,6 +356,9 @@ Json::Value Control::SelectStatusRecord(Json::Value &queryjson) {
 
 // ----------------------判题----------------------------
 Json::Value Control::GetJudgeCode(Json::Value judgejson) {
+    // 如果不是用户，无权限
+    if (!UserList::GetInstance()->IsOrdinaryUser(judgejson))
+        return NoPermission;
     Json::Value resjson;
     // 传入Json(ProblemId,UserId,UserNickName,Code,Language,TimeLimit,MemoryLimit,JudgeNum,ProblemTitle)
 
@@ -337,10 +421,23 @@ Json::Value Control::GetJudgeCode(Json::Value judgejson) {
     return resjson;
 }
 
+// ----------------------token----------------------------
+// 登录用户通过Token
+Json::Value Control::LoginUserByToken(Json::Value &loginjson) {
+    return UserList::GetInstance()->LoginUserByToken(loginjson);
+}
+
 // 构造函数
 Control::Control() {
     // 初始化题目标签
     Tag::GetInstance()->InitProblemTags();
+
+    // 初始化用户权限
+    UserList::GetInstance()->InitUserAuthority();
+
+    // 初始化返回变量
+    NoPermission["Result"] = "401";
+    NoPermission["Reason"] = "无权限";
 }
 
 // 析构函数
